refacto: Motidy projects code structure (clean / reorganize)
