+++ src/mock/model.rs
@@ -1,6 +1,5 @@
 use serde::{Deserialize, Serialize};
 use std::cmp::PartialEq;
-use std::fmt::{Display, Formatter};
 
 #[derive(Debug)]
 pub struct IncomingRequest {
@@ -65,34 +64,10 @@ impl Expectation {
     }
 }
 
-impl Display for Expectation {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
-        write!(
-            f,
-            "Expectation {{ \n\tRequest {} {} \n\tResponse {} {} \n\t\t{} \n}}",
-            self.request
-                .method
-                .as_ref()
-                .map(|method| format!("{:?}", method))
-                .unwrap_or("_".to_string()),
-            self.request.path.as_ref().unwrap_or(&"_".to_string()),
-            self.response
-                .status_code
-                .map(|method| format!("{}", method))
-                .unwrap_or("_".to_string()),
-            self.response
-                .status_reason
-                .as_ref()
-                .unwrap_or(&"_".to_string()),
-            self.response.body.as_ref().unwrap_or(&"_".to_string())
-        )
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::expectation::model::{HttpRequest, HttpResponse};
+    use crate::mock::model::{HttpRequest, HttpResponse};
 
     #[test]
     fn test_deserialize_http_method() {
